+++ src/app/[locale]/games/page.tsx
@@ -1,7 +1,9 @@
 import GamesHub from "@/components/pages/games-hub";
+import { getGames } from "@/lib/helpers/data";
 
-export default function Games(){
+export default async function Games(){
+     const games = await getGames()
      return (
-          <GamesHub/>
+          <GamesHub games={games}/>
      )
 }