+++ src/components/pages/games-hub.tsx
@@ -2,9 +2,13 @@
 import PageLayout from "@/components/layout";
 import GamesSection from "@/components/sections/games";
 import HeroSection from "@/components/sections/hero";
+import { ICard } from "@/lib/types";
 import { useTranslations } from "next-intl";
 
-export default function GamesHub(){
+interface GamesHubProps{
+     games: ICard<"game">[]
+}
+export default function GamesHub({games}: GamesHubProps){
      const t = useTranslations("games")
      return (
           <PageLayout>
@@ -14,7 +18,7 @@ export default function GamesHub(){
                     link="#main-games"
                     linkText={t("startPlaying")}
                />
-               <GamesSection/>
+               <GamesSection games={games}/>
           </PageLayout>
      )
 }