+++ changelog.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## 2.0.2 - Unreleased
 ### Improved
 - **Downloads section** - Moved Downloads Data into static resources
+- **Games section** and **Christmas games section** - Moved games list into static resources
 
 ## 2.0.1 - 2026-08-30
 ### Added